fix: totalCount 데이터 타입 Long으로 변경

+++ application/src/main/java/org/depromeet/spot/application/review/dto/response/ReviewListResponse.java
@@ -8,7 +8,7 @@ import org.depromeet.spot.domain.review.ReviewListResult;
 public record ReviewListResponse(
         List<KeywordCountResponse> keywords,
         List<ReviewResponse> reviews,
-        int totalCount,
+        Long totalCount,
         int filteredCount,
         int offset,
         int limit,
